Adds colorization for RAML
The RAML language has a handful of issues and I haven't yet been able to
identify all color combinations, but this is a good start at the majority
of useful scopes.

Some of these depend on a couple of new scopes that have not yet been
accepted upstream in the RAML plugin repository but adding them to the
Neon theme won't break anything in the meantime.
(see: https://github.com/mulesoft/raml-sublime-plugin/pull/7)